+++ .gitignore
@@ -13,6 +13,9 @@ src/gevent/greenlet.c
 src/gevent/_ident.c
 src/gevent/_imap.c
 src/gevent/event.c
+src/gevent/_hub_local.c
+src/gevent/_waiter.c
+src/gevent/queue.c
 src/gevent/libev/corecext.c
 src/gevent/libev/corecext.h
 src/gevent/libev/_corecffi.c
