+++ CHANGES.rst
@@ -50,7 +50,8 @@ Enhancements
   reducing the overhead of ``[Thread]Pool.imap``.
 
 - The classes `gevent.event.Event` and `gevent.event.AsyncResult`
-  are compiled with Cython for improved performance. Please report any
+  are compiled with Cython for improved performance, as is the
+  ``gevent.queue`` module and ``gevent.hub.Waiter``. Please report any
   compatibility issues.
 
 Monitoring and Debugging
