+++ src/gevent/queue.py
@@ -1,4 +1,6 @@
 # Copyright (c) 2009-2012 Denis Bilenko. See LICENSE for details.
+# copyright (c) 2018 gevent
+# cython: auto_pickle=False,embedsignature=True,always_allow_keywords=False
 """
 Synchronized queues.
 
@@ -27,7 +29,9 @@ class, not an instance or subclass).
 
 from __future__ import absolute_import
 import sys
-import heapq
+from heapq import heappush as _heappush
+from heapq import heappop as _heappop
+from heapq import heapify as _heapify
 import collections
 
 if sys.version_info[0] == 2:
@@ -38,10 +42,9 @@ Full = __queue__.Full
 Empty = __queue__.Empty
 
 from gevent.timeout import Timeout
-from gevent.hub import _get_hub_noargs as get_hub
-from gevent.hub import Waiter
-from gevent.hub import getcurrent
-from gevent.hub import InvalidSwitchError
+from gevent._hub_local import get_hub_noargs as get_hub
+from greenlet import getcurrent
+from gevent.exceptions import InvalidSwitchError
 
 
 __implements__ = ['Queue', 'PriorityQueue', 'LifoQueue']
@@ -61,12 +64,18 @@ def _safe_remove(deq, item):
     except ValueError:
         pass
 
+import gevent._waiter
+locals()['Waiter'] = gevent._waiter.Waiter
 
-class ItemWaiter(Waiter):
-    __slots__ = ['item', 'queue']
+class ItemWaiter(Waiter): # pylint:disable=undefined-variable
+    # pylint:disable=assigning-non-slot
+    __slots__ = (
+        'item',
+        'queue',
+    )
 
     def __init__(self, item, queue):
-        Waiter.__init__(self)
+        Waiter.__init__(self) # pylint:disable=undefined-variable
         self.item = item
         self.queue = queue
 
@@ -97,19 +106,26 @@ class Queue(object):
        previously anyway, but that wasn't the case for PyPy.
     """
 
-    _warn_depth = 2
+    __slots__ = (
+        '_maxsize',
+        'getters',
+        'putters',
+        'hub',
+        '_event_unlock',
+        'queue',
+    )
 
-    def __init__(self, maxsize=None, items=None):
+    def __init__(self, maxsize=None, items=(), _warn_depth=2):
         if maxsize is not None and maxsize <= 0:
-            self.maxsize = None
             if maxsize == 0:
                 import warnings
                 warnings.warn(
                     'Queue(0) now equivalent to Queue(None); if you want a channel, use Channel',
                     DeprecationWarning,
-                    stacklevel=self._warn_depth)
-        else:
-            self.maxsize = maxsize
+                    stacklevel=_warn_depth)
+            maxsize = None
+
+        self._maxsize = maxsize if maxsize is not None else -1
         # Explicitly maintain order for getters and putters that block
         # so that callers can consistently rely on getting things out
         # in the apparent order they went in. This was once required by
@@ -126,23 +142,25 @@ class Queue(object):
         self.putters = collections.deque()
         self.hub = get_hub()
         self._event_unlock = None
-        if items:
-            self._init(maxsize, items)
-        else:
-            self._init(maxsize)
+        self.queue = self._create_queue(items)
 
-    # QQQ make maxsize into a property with setter that schedules unlock if necessary
+    @property
+    def maxsize(self):
+        return self._maxsize if self._maxsize > 0 else None
+
+    @maxsize.setter
+    def maxsize(self, nv):
+        # QQQ make maxsize into a property with setter that schedules unlock if necessary
+        if nv is None or nv <= 0:
+            self._maxsize = -1
+        else:
+            self._maxsize = nv
 
     def copy(self):
         return type(self)(self.maxsize, self.queue)
 
-    def _init(self, maxsize, items=None):
-        # FIXME: Why is maxsize unused or even passed?
-        # pylint:disable=unused-argument
-        if items:
-            self.queue = collections.deque(items)
-        else:
-            self.queue = collections.deque()
+    def _create_queue(self, items=()):
+        return collections.deque(items)
 
     def _get(self):
         return self.queue.popleft()
@@ -198,7 +216,12 @@ class Queue(object):
            to return True for backwards compatibility.
         """
         return True
-    __nonzero__ = __bool__
+
+    def __nonzero__(self):
+        # Py2.
+        # For Cython; __bool__ becomes a special method that we can't
+        # get by name.
+        return True
 
     def empty(self):
         """Return ``True`` if the queue is empty, ``False`` otherwise."""
@@ -209,7 +232,7 @@ class Queue(object):
 
         ``Queue(None)`` is never full.
         """
-        return self.maxsize is not None and self.qsize() >= self.maxsize
+        return self._maxsize > 0 and self.qsize() >= self._maxsize
 
     def put(self, item, block=True, timeout=None):
         """Put an item into the queue.
@@ -222,7 +245,7 @@ class Queue(object):
         is immediately available, else raise the :class:`Full` exception (*timeout*
         is ignored in that case).
         """
-        if self.maxsize is None or self.qsize() < self.maxsize:
+        if self._maxsize == -1 or self.qsize() < self._maxsize:
             # there's a free slot, put an item right away
             self._put(item)
             if self.getters:
@@ -230,10 +253,10 @@ class Queue(object):
         elif self.hub is getcurrent():
             # We're in the mainloop, so we cannot wait; we can switch to other greenlets though.
             # Check if possible to get a free slot in the queue.
-            while self.getters and self.qsize() and self.qsize() >= self.maxsize:
+            while self.getters and self.qsize() and self.qsize() >= self._maxsize:
                 getter = self.getters.popleft()
                 getter.switch(getter)
-            if self.qsize() < self.maxsize:
+            if self.qsize() < self._maxsize:
                 self._put(item)
                 return
             raise Full
@@ -283,7 +306,7 @@ class Queue(object):
             # to return. No choice...
             raise Empty()
 
-        waiter = Waiter()
+        waiter = Waiter() # pylint:disable=undefined-variable
         timeout = Timeout._start_new_or_dummy(timeout, Empty)
         try:
             self.getters.append(waiter)
@@ -349,7 +372,7 @@ class Queue(object):
     def _unlock(self):
         while True:
             repeat = False
-            if self.putters and (self.maxsize is None or self.qsize() < self.maxsize):
+            if self.putters and (self._maxsize == -1 or self.qsize() < self._maxsize):
                 repeat = True
                 try:
                     putter = self.putters.popleft()
@@ -372,16 +395,31 @@ class Queue(object):
     def __iter__(self):
         return self
 
-    def next(self):
+    def __next__(self):
         result = self.get()
         if result is StopIteration:
             raise result
         return result
 
-    __next__ = next
+    next = __next__ # Py2
 
 
+class UnboundQueue(Queue):
+    # A specialization of Queue that knows it can never
+    # be bound. Changing its maxsize has no effect.
 
+    __slots__ = ()
+
+    def __init__(self, maxsize=None, items=()):
+        if maxsize is not None:
+            raise ValueError("UnboundQueue has no maxsize")
+        Queue.__init__(self, maxsize, items)
+        self.putters = None # Will never be used.
+
+    def put(self, item, block=True, timeout=None):
+        self._put(item)
+        if self.getters:
+            self._schedule_unlock()
 
 
 class PriorityQueue(Queue):
@@ -395,30 +433,27 @@ class PriorityQueue(Queue):
        Previously it was just assumed that they were already a heap.
     '''
 
-    def _init(self, maxsize, items=None):
-        if items:
-            self.queue = list(items)
-            heapq.heapify(self.queue)
-        else:
-            self.queue = []
+    __slots__ = ()
 
-    def _put(self, item, heappush=heapq.heappush):
-        # pylint:disable=arguments-differ
-        heappush(self.queue, item)
+    def _create_queue(self, items=()):
+        q = list(items)
+        _heapify(q)
+        return q
 
-    def _get(self, heappop=heapq.heappop):
-        # pylint:disable=arguments-differ
-        return heappop(self.queue)
+    def _put(self, item):
+        _heappush(self.queue, item)
+
+    def _get(self):
+        return _heappop(self.queue)
 
 
 class LifoQueue(Queue):
     '''A subclass of :class:`Queue` that retrieves most recently added entries first.'''
 
-    def _init(self, maxsize, items=None):
-        if items:
-            self.queue = list(items)
-        else:
-            self.queue = []
+    __slots__ = ()
+
+    def _create_queue(self, items=()):
+        return list(items)
 
     def _put(self, item):
         self.queue.append(item)
@@ -436,9 +471,12 @@ class JoinableQueue(Queue):
     :meth:`task_done` and :meth:`join` methods.
     """
 
-    _warn_depth = 3
+    __slots__ = (
+        '_cond',
+        'unfinished_tasks',
+    )
 
-    def __init__(self, maxsize=None, items=None, unfinished_tasks=None):
+    def __init__(self, maxsize=None, items=(), unfinished_tasks=None):
         """
 
         .. versionchanged:: 1.1a1
@@ -446,8 +484,9 @@ class JoinableQueue(Queue):
            (if any) will be considered unfinished.
 
         """
+        Queue.__init__(self, maxsize, items, _warn_depth=3)
+
         from gevent.event import Event
-        Queue.__init__(self, maxsize, items)
         self._cond = Event()
         self._cond.set()
 
@@ -514,6 +553,13 @@ class JoinableQueue(Queue):
 
 class Channel(object):
 
+    __slots__ = (
+        'getters',
+        'putters',
+        'hub',
+        '_event_unlock',
+    )
+
     def __init__(self, maxsize=1):
         # We take maxsize to simplify certain kinds of code
         if maxsize != 1:
@@ -561,7 +607,7 @@ class Channel(object):
         if not block:
             timeout = 0
 
-        waiter = Waiter()
+        waiter = Waiter() # pylint:disable=undefined-variable
         item = (item, waiter)
         self.putters.append(item)
         timeout = Timeout._start_new_or_dummy(timeout, Full)
@@ -590,7 +636,7 @@ class Channel(object):
         if not block:
             timeout = 0
 
-        waiter = Waiter()
+        waiter = Waiter() # pylint:disable=undefined-variable
         timeout = Timeout._start_new_or_dummy(timeout, Empty)
         try:
             self.getters.append(waiter)
@@ -620,10 +666,13 @@ class Channel(object):
     def __iter__(self):
         return self
 
-    def next(self):
+    def __next__(self):
         result = self.get()
         if result is StopIteration:
             raise result
         return result
 
-    __next__ = next # py3
+    next = __next__ # Py2
+
+from gevent._util import import_c_accel
+import_c_accel(globals(), 'gevent._queue')
